app/requirements.txt
bigframes
google-cloud-aiplatform
google-cloud-pubsub
google-cloud-storage
gunicorn
json-repair
langchain
langgraph
langchain_google_vertexai
mesop
vertexai
